chore: error boundary box 처리
