+++ src/components/index.ts
@@ -6,7 +6,6 @@ export { default as Drawer } from './Drawer/Drawer';
 export { default as Dropdown } from './Dropdown/Dropdown';
 export { default as AuthErrorFallback } from './Error/AuthErrorFallback';
 export { default as CountBoxFallback } from './Error/CountBoxFallback';
-export { default as ErrorBox } from './Error/ErrorBox';
 export { default as ErrorResetBoundary } from './Error/ErrorResetBoundary';
 export { default as GlobalErrorBoundary } from './Error/GlobalErrorBoundary';
 export { default as GlobalErrorFallbackSection } from './Error/GlobalErrorFallbackSection';
@@ -37,6 +36,8 @@ export { default as UploadButton } from './UploadButton/UploadButton';
 export * from './Button/BackButton';
 export * from './Button/ToggleButton';
 export * from './Deferred/DeferredComponent';
+export * from './Error/ErrorBox';
+export * from './Error/ErrorContent';
 export * from './Footer/Footer';
 export * from './FormField/ControlledFormField';
 export * from './Input/Input';
