+++ src/pages/profile/components/GridSectionCard.tsx
@@ -1,3 +1,4 @@
+import { Flex } from '@/components';
 import { boxShadow } from '@/styles/common';
 import { styled } from '@mui/material';
 import { PropsWithChildren } from 'react';
@@ -7,21 +8,27 @@ interface Props extends PropsWithChildren {
 }
 
 export const GridSectionCard = ({ children, onClick }: Props) => {
-  return <S.Section onClick={onClick}>{children}</S.Section>;
+  return (
+    <S.Section
+      width="100%"
+      height="250px"
+      justify="center"
+      align="center"
+      padding="2rem"
+      onClick={onClick}
+    >
+      {children}
+    </S.Section>
+  );
 };
 
 const S = {
-  Section: styled('div')`
-    align-items: center;
+  Section: styled(Flex.Column)`
     background-color: ${({ theme }) => theme.palette.variant.default};
     border-radius: 1rem;
     cursor: pointer;
-    height: 250px;
-    justify-content: center;
-    padding: 2rem;
     ${boxShadow};
     position: relative;
-    width: 100%;
 
     &:hover,
     &:active {
